# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_searchUsers",
        "parameters": [
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "emailVerified",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "country",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "city",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdAfter",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdBefore",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/activate": {
      "put": {
        "operationId": "UsersController_activateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/deactivate": {
      "put": {
        "operationId": "UsersController_deactivateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}/verify-email": {
      "put": {
        "operationId": "UsersController_verifyUserEmail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/orders/user/{userId}": {
      "get": {
        "description": "获取指定用户的订单列表",
        "operationId": "OrdersController_findByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "用户ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "订单状态筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "排序方向",
            "schema": {
              "example": "DESC",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResponseDto"
                            },
                            {
                              "properties": {
                                "items": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/Object"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "查询参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "未授权访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "获取用户订单",
        "tags": [
          "订单管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "Caddy Style Shopping API",
    "description": "\n        ## 电商系统 API 文档\n        \n        ### 功能模块\n        - **认证管理**: 用户注册、登录、JWT令牌管理\n        - **用户管理**: 用户信息、权限管理、地址管理\n        - **商品管理**: 商品CRUD、分类、库存管理\n        - **购物车管理**: 购物车操作、商品选择、批量操作\n        - **订单管理**: 订单创建、支付、状态跟踪\n        - **支付管理**: 多种支付方式、支付回调处理\n        - **系统管理**: 缓存管理、性能监控、健康检查\n        \n        ### 认证方式\n        使用 JWT Bearer Token 进行认证，请在请求头中添加：\n        `Authorization: Bearer <your-token>`\n        \n        ### 响应格式\n        所有API响应都遵循统一格式：\n        ```json\n        {\n          \"code\": 200,\n          \"message\": \"操作成功\",\n          \"data\": {},\n          \"timestamp\": \"2025-01-26T10:30:00Z\",\n          \"requestId\": \"req_123456789\"\n        }\n        ```\n        \n        ### 错误码说明\n        - **1xxx**: 通用错误（参数错误、资源不存在等）\n        - **2xxx**: 认证错误（未授权、令牌过期等）\n        - **3xxx**: 业务错误（库存不足、订单状态错误等）\n        - **5xxx**: 系统错误（服务器错误、数据库错误等）\n      ",
    "version": "1.0.0",
    "contact": {
      "name": "开发团队",
      "url": "https://github.com/your-org/caddy-style-shopping",
      "email": "dev@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发环境"
    },
    {
      "url": "https://api-staging.example.com",
      "description": "测试环境"
    },
    {
      "url": "https://api.example.com",
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "错误状态码",
            "example": 400
          },
          "message": {
            "type": "string",
            "description": "错误消息",
            "example": "请求参数错误"
          },
          "errors": {
            "description": "错误详情",
            "example": [
              "用户名不能为空",
              "密码长度至少6位"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timestamp": {
            "type": "string",
            "description": "时间戳",
            "example": "2025-01-26T10:30:00Z"
          },
          "path": {
            "type": "string",
            "description": "请求路径",
            "example": "/api/auth/login"
          },
          "requestId": {
            "type": "string",
            "description": "请求ID",
            "example": "req_123456789"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "object"
          },
          "preferences": {
            "type": "object"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "object"
          },
          "preferences": {
            "type": "object"
          }
        }
      },
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "响应状态码",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "timestamp": {
            "type": "string",
            "description": "时间戳",
            "example": "2025-01-26T10:30:00Z"
          },
          "requestId": {
            "type": "string",
            "description": "请求ID",
            "example": "req_123456789"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp"
        ]
      },
...
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "数据列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "total": {
            "type": "number",
            "description": "总数量",
            "example": 100
          },
          "page": {
            "type": "number",
            "description": "当前页码",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "每页数量",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "description": "总页数",
            "example": 5
          },
          "hasNext": {
            "type": "boolean",
            "description": "是否有下一页",
            "example": true
          },
          "hasPrev": {
            "type": "boolean",
            "description": "是否有上一页",
            "example": false
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "totalPages",
          "hasNext",
          "hasPrev"
        ]
      }
    }
  }
}